UPD | 패키지 구조에 따른 변경사항

+++ src/main/kotlin/commerce/hosinsa/domain/dto/coupon/ResponseCoupon.kt
@@ -5,7 +5,7 @@ import java.time.LocalDateTime
 
 class ResponseCoupon @QueryProjection constructor(
     val couponName: String,
-    val discount: Short,
+    val discount: Byte,
     val brandName: String?,
     val assignedAt: LocalDateTime,
     val expiredAt: LocalDateTime,
